+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [0.1.0](https://github.com/sandovalrr/project-context/compare/v0.0.0...v0.1.0) (2026-07-18)
 
+### Bug Fixes
+
+- generate SBOM from Bun installs ([1b98eed](https://github.com/sandovalrr/project-context/commit/1b98eed681c215b561b6a665bfea6a18a050bdfb))
+
+### Features
+
+- add client-ready MCP manifests ([6833165](https://github.com/sandovalrr/project-context/commit/68331656ab8d5e5b5864b8486e3a0784844392e0))
+
 # Changelog
 
 All notable changes to this project are documented here. Releases follow